[ESP32] decouple Thread support from specific chip model
This change makes OpenThread available on all ESP32 series when
CONFIG_OPENTHREAD_ENABLED is on.
