+++ examples/all-clusters-app/esp32/main/CMakeLists.txt
@@ -107,6 +107,10 @@ if("${CONFIG_DEVICE_TYPE_ESP32_C3_DEVKITM}" STREQUAL "y")
     list(APPEND PRIV_REQUIRES_LIST led_strip)
 endif()
 
+if (CONFIG_OPENTHREAD_ENABLED)
+    list(APPEND PRIV_REQUIRES_LIST openthread)
+endif()
+
 idf_component_register(PRIV_INCLUDE_DIRS ${PRIV_INCLUDE_DIRS_LIST}
                        SRC_DIRS ${SRC_DIRS_LIST}
                        PRIV_REQUIRES ${PRIV_REQUIRES_LIST})
