+++ examples/all-clusters-app/esp32/partitions.csv
@@ -2,5 +2,5 @@
 # Note: if you have increased the bootloader size, make sure to update the offsets to avoid overlap
 nvs,      data, nvs,     ,        0x6000,
 phy_init, data, phy,     ,        0x1000,
-# Factory partition size about 1.9MB
-factory,  app,  factory, ,        1945K,
+factory,  app,  factory, ,        1900K,
+ot_storage, data, 0x3a,            , 0x2000,
