+++ src/platform/ESP32/BUILD.gn
@@ -26,6 +26,7 @@ static_library("ESP32") {
     "CHIPDevicePlatformEvent.h",
     "ConfigurationManagerImpl.cpp",
     "ConfigurationManagerImpl.h",
+    "ConnectivityManagerImpl.cpp",
     "ConnectivityManagerImpl.h",
     "DeviceNetworkProvisioningDelegateImpl.cpp",
     "DeviceNetworkProvisioningDelegateImpl.h",
@@ -51,23 +52,21 @@ static_library("ESP32") {
     "${chip_root}/src/crypto",
     "${chip_root}/src/platform:platform_base",
   ]
+
+  if (chip_enable_wifi) {
+    sources += [
+      "ConnectivityManagerImpl_WiFi.cpp",
+      "WiFiProvisioning.cpp",
+      "WiFiProvisioning.h",
+    ]
+  }
+
   if (chip_enable_openthread) {
     sources += [
-      "../OpenThread/MdnsImpl.cpp",
       "../OpenThread/OpenThreadUtils.cpp",
-      "ConnectivityManagerImpl_Thread.cpp",
-      "ConnectivityManagerImpl_Thread.h",
       "ESPThreadConfig.h",
       "ThreadStackManagerImpl.cpp",
       "ThreadStackManagerImpl.h",
-    ]
-  } else {
-    sources += [
-      "ConnectivityManagerImpl_WiFi.cpp",
-      "ConnectivityManagerImpl_WiFi.h",
-      "MdnsImpl.cpp",
-      "ServiceProvisioning.cpp",
-      "ServiceProvisioning.h",
     ]
   }
 }
