--- src/platform/ESP32/ConnectivityManagerImpl_Thread.h
@@ -1,94 +1,0 @@
-/*
- *
- *    Copyright (c) 2020-2021 Project CHIP Authors
- *    Copyright (c) 2018 Nest Labs, Inc.
- *    All rights reserved.
- *
- *    Licensed under the Apache License, Version 2.0 (the "License");
- *    you may not use this file except in compliance with the License.
- *    You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
- *    Unless required by applicable law or agreed to in writing, software
- *    distributed under the License is distributed on an "AS IS" BASIS,
- *    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *    See the License for the specific language governing permissions and
- *    limitations under the License.
- */
-
-#pragma once
-
-#include <lib/support/logging/CHIPLogging.h>
-#include <platform/ConnectivityManager.h>
-#include <platform/internal/GenericConnectivityManagerImpl.h>
-#include <platform/internal/GenericConnectivityManagerImpl_NoWiFi.h>
-#if CHIP_DEVICE_CONFIG_ENABLE_CHIPOBLE
-#include <platform/internal/GenericConnectivityManagerImpl_BLE.h>
-#else
-#include <platform/internal/GenericConnectivityManagerImpl_NoBLE.h>
-#endif
-#include <platform/internal/GenericConnectivityManagerImpl_Thread.h>
-
-namespace Inet {
-class IPAddress;
-} // namespace Inet
-
-namespace chip {
-namespace DeviceLayer {
-
-/**
- * Concrete implementation of the ConnectivityManager singleton object for the ESP32 platform.
- */
-class ConnectivityManagerImpl final : public ConnectivityManager,
-                                      public Internal::GenericConnectivityManagerImpl_NoWiFi<ConnectivityManagerImpl>,
-#if CHIP_DEVICE_CONFIG_ENABLE_CHIPOBLE
-                                      public Internal::GenericConnectivityManagerImpl_BLE<ConnectivityManagerImpl>,
-#else
-                                      public Internal::GenericConnectivityManagerImpl_NoBLE<ConnectivityManagerImpl>,
-#endif
-                                      public Internal::GenericConnectivityManagerImpl_Thread<ConnectivityManagerImpl>,
-                                      public Internal::GenericConnectivityManagerImpl<ConnectivityManagerImpl>
-{
-
-    // Allow the ConnectivityManager interface class to delegate method calls to
-    // the implementation methods provided by this class.
-    friend class ConnectivityManager;
-
-private:
-    // ===== Members that implement the ConnectivityManager abstract interface.
-
-    CHIP_ERROR _Init(void);
-    void _OnPlatformEvent(const ChipDeviceEvent * event);
-    // ===== Members for internal use by the following friends.
-
-    friend ConnectivityManager & ConnectivityMgr(void);
-    friend ConnectivityManagerImpl & ConnectivityMgrImpl(void);
-
-    static ConnectivityManagerImpl sInstance;
-};
-
-/**
- * Returns the public interface of the ConnectivityManager singleton object.
- *
- * Chip applications should use this to access features of the ConnectivityManager object
- * that are common to all platforms.
- */
-inline ConnectivityManager & ConnectivityMgr(void)
-{
-    return ConnectivityManagerImpl::sInstance;
-}
-
-/**
- * Returns the platform-specific implementation of the ConnectivityManager singleton object.
- *
- * Chip applications can use this to gain access to features of the ConnectivityManager
- * that are specific to the ESP32 platform.
- */
-inline ConnectivityManagerImpl & ConnectivityMgrImpl(void)
-{
-    return ConnectivityManagerImpl::sInstance;
-}
-
-} // namespace DeviceLayer
-} // namespace chip
