+++ src/platform/ESP32/ConnectivityManagerImpl_WiFi.cpp
@@ -20,10 +20,6 @@
 #include <platform/internal/CHIPDeviceLayerInternal.h>
 
 #include <platform/ConnectivityManager.h>
-#if CHIP_DEVICE_CONFIG_ENABLE_CHIPOBLE
-#include <platform/internal/GenericConnectivityManagerImpl_BLE.cpp>
-#endif
-#include <platform/internal/GenericConnectivityManagerImpl_WiFi.cpp>
 
 #include <lib/support/CodeUtils.h>
 #include <lib/support/logging/CHIPLogging.h>
@@ -39,15 +35,7 @@
 #include <lwip/nd6.h>
 #include <lwip/netif.h>
 
-#include <type_traits>
-
-#if !CHIP_DEVICE_CONFIG_ENABLE_WIFI_STATION
-#error "WiFi Station support must be enabled when building for ESP32"
-#endif
-
-#if !CHIP_DEVICE_CONFIG_ENABLE_WIFI_AP
-#error "WiFi AP support must be enabled when building for ESP32"
-#endif
+#if CHIP_DEVICE_CONFIG_ENABLE_WIFI
 
 using namespace ::chip;
 using namespace ::chip::Inet;
@@ -57,8 +45,6 @@ using chip::DeviceLayer::Internal::ESP32Utils;
 
 namespace chip {
 namespace DeviceLayer {
-
-ConnectivityManagerImpl ConnectivityManagerImpl::sInstance;
 
 ConnectivityManager::WiFiStationMode ConnectivityManagerImpl::_GetWiFiStationMode(void)
 {
@@ -387,9 +373,7 @@ CHIP_ERROR ConnectivityManagerImpl::_GetAndLogWifiStatsCounters(void)
     return CHIP_NO_ERROR;
 }
 
-// ==================== ConnectivityManager Platform Internal Methods ====================
-
-CHIP_ERROR ConnectivityManagerImpl::_Init()
+CHIP_ERROR ConnectivityManagerImpl::InitWiFi()
 {
     mLastStationConnectFailTime     = 0;
     mLastAPDemandTime               = 0;
@@ -448,7 +432,7 @@ CHIP_ERROR ConnectivityManagerImpl::_Init()
     return CHIP_NO_ERROR;
 }
 
-void ConnectivityManagerImpl::_OnPlatformEvent(const ChipDeviceEvent * event)
+void ConnectivityManagerImpl::OnWiFiPlatformEvent(const ChipDeviceEvent * event)
 {
     // Handle ESP system events...
     if (event->Type == DeviceEventType::kESPSystemEvent)
@@ -535,8 +519,6 @@ void ConnectivityManagerImpl::_OnWiFiStationProvisionChange()
     // Schedule a call to the DriveStationState method to adjust the station state as needed.
     DeviceLayer::SystemLayer().ScheduleWork(DriveStationState, NULL);
 }
-
-// ==================== ConnectivityManager Private Methods ====================
 
 void ConnectivityManagerImpl::DriveStationState()
 {
@@ -974,8 +956,6 @@ void ConnectivityManagerImpl::OnStationIPv4AddressAvailable(const ip_event_got_i
     }
 #endif // CHIP_PROGRESS_LOGGING
 
-    RefreshMessageLayer();
-
     UpdateInternetConnectivityState();
 
     ChipDeviceEvent event;
@@ -987,8 +967,6 @@ void ConnectivityManagerImpl::OnStationIPv4AddressAvailable(const ip_event_got_i
 void ConnectivityManagerImpl::OnStationIPv4AddressLost(void)
 {
     ChipLogProgress(DeviceLayer, "IPv4 address lost on WiFi station interface");
-
-    RefreshMessageLayer();
 
     UpdateInternetConnectivityState();
 
@@ -1007,8 +985,6 @@ void ConnectivityManagerImpl::OnIPv6AddressAvailable(const ip_event_got_ip6_t & 
     }
 #endif // CHIP_PROGRESS_LOGGING
 
-    RefreshMessageLayer();
-
     UpdateInternetConnectivityState();
 
     ChipDeviceEvent event;
@@ -1017,7 +993,7 @@ void ConnectivityManagerImpl::OnIPv6AddressAvailable(const ip_event_got_ip6_t & 
     PlatformMgr().PostEventOrDie(&event);
 }
 
-void ConnectivityManagerImpl::RefreshMessageLayer(void) {}
-
 } // namespace DeviceLayer
 } // namespace chip
+
+#endif // CHIP_DEVICE_CONFIG_ENABLE_WIFI
